+++ lib/src/commands/protos.dart
@@ -5,18 +5,23 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:intl/intl.dart';
 import 'package:path/path.dart';
 import 'package:upcode_ci/src/commands/command.dart';
 
 class ProtosCommand extends UpcodeCommand {
-  ProtosCommand(Map<String, dynamic> config) : super(config);
+  ProtosCommand(Map<String, dynamic> config) : super(config) {
+    argParser
+      ..addFlag('all', defaultsTo: false, help: 'Build all files.')
+      ..addFlag('js', defaultsTo: false, help: 'Build js proto files.')
+      ..addFlag('dart', defaultsTo: false, help: 'Build dart proto files.')
+      ..addFlag('descriptor', defaultsTo: false, help: 'Generate API descriptor for Cloud Endpoints.');
+  }
 
   @override
   final String name = 'protos';
 
   @override
-  final String description = 'Generate implementation files in dart from proto files.';
+  final String description = 'Generate implementation files in dart and js, and the API descriptor from proto files.';
 
   void _deleteCurrent(List<String> dirs) {
     for (String dir in dirs) {
@@ -36,35 +41,41 @@ class ProtosCommand extends UpcodeCommand {
       .map((element) => element.path)
       .toList();
 
-  List<String> get _generatedDartFiles => Directory(dartProtoDir)
-      .listSync(recursive: true, followLinks: false)
-      .whereType<File>()
-      .where((File file) => file.path.endsWith('.dart'))
-      .map((File file) => file.path.split(join('lib', 'generated'))[1])
-      .where((String path) => !path.contains('${context.separator}struct.'))
-      .map((String path) => "export '${path.split(context.separator).skip(1).join('/')}';")
-      .toList()
-        ..sort();
-
-  void _buildDartImports() {
-    File(join(flutterGeneratedDir, 'protos.dart')).writeAsStringSync('''// GENERATED FILE, DO NOT EDIT
-// Last update ${DateFormat.yMMMMd().add_Hm().format(DateTime.now().toUtc())}
-
-library protos;
-
-${_generatedDartFiles.join('\n')}
-''');
-  }
-
   @override
   FutureOr<dynamic> run() async {
     if (!flutterGeneratedDir.dir.existsSync()) {
       flutterGeneratedDir.dir.createSync(recursive: true);
     }
 
-    final List<String> dirsToDelete = <String>[dartProtoDir];
+    bool buildJs;
+    bool buildDart;
+    bool descriptor;
+    if (!argResults.wasParsed('js') && !argResults.wasParsed('dart') && !argResults.wasParsed('descriptor')) {
+      buildJs = true;
+      buildDart = true;
+      descriptor = true;
+    } else {
+      final bool all = argResults['all'];
+      buildJs = argResults['js'] || all;
+      buildDart = argResults['dart'] || all;
+      descriptor = argResults['descriptor'] || all;
+    }
+
+    if (!buildJs && !buildDart && !descriptor) {
+      stdout.writeln('Nothing to build.');
+      return;
+    }
+
+    final List<String> dirsToDelete = [if (buildDart) dartProtoDir, if (buildJs) protoApiOutDir];
     if (dirsToDelete.isNotEmpty) {
       execute(() => _deleteCurrent(dirsToDelete), 'Delete existing proto implementation');
+    }
+
+    if (!protoApiOutDir.existsSync()) {
+      await Directory(protoApiOutDir).createSync(recursive: true);
+    }
+    if (!dartProtoDir.existsSync()) {
+      await Directory(dartProtoDir).createSync(recursive: true);
     }
 
     await execute(
@@ -72,16 +83,30 @@ ${_generatedDartFiles.join('\n')}
         return runCommand(
           'protoc',
           <String>[
-            '--dart_out=grpc:$dartProtoDir',
+            if (buildJs) ...<String>[
+              '--js_out=import_style=commonjs,binary:$protoApiOutDir',
+              '--ts_out=$protoApiOutDir',
+              '--grpc_out=$protoApiOutDir',
+              '--plugin=protoc-gen-grpc=${join(apiDir, 'node_modules', '.bin', 'grpc_tools_node_protoc_plugin')}',
+              '--plugin=protoc-gen-ts=${join(apiDir, 'node_modules', '.bin', 'protoc-gen-ts')}',
+            ],
+            if (buildDart) '--dart_out=grpc:$dartProtoDir',
+            if (descriptor) ...<String>[
+              '--include_imports',
+              '--include_source_info',
+              '--descriptor_set_out=$apiDescriptor',
+            ],
             '--proto_path=$protoSrcDir',
-            ..._protoFiles,
+            ..._protoFiles
           ],
-          workingDirectory: pwd,
+          workingDirectory: Directory.current.path,
         );
       },
-      'Proto building dart implementation',
+      'Proto building: ${[
+        if (buildJs) 'js implementation',
+        if (buildDart) 'dart implementation',
+        if (descriptor) 'api descriptor',
+      ].join(', ')}',
     );
-
-    execute(_buildDartImports, 'Build Dart import file');
   }
 }